Move common helper functions
