+++ eclair-core/src/test/scala/fr/acinq/eclair/Pipe.scala
@@ -17,7 +17,7 @@
 package fr.acinq.eclair
 
 import akka.actor.{Actor, ActorLogging, ActorRef, Stash}
-import fr.acinq.eclair.channel.Commitments.msg2String
+import fr.acinq.eclair.channel.Common.msg2String
 import fr.acinq.eclair.wire.protocol.LightningMessage
 
 /**
